requirements.txt
flask
google-genai
pillow
werkzeug
python-dotenv